+++ src/utils/middleware.go
@@ -346,6 +346,17 @@ func EnsureHostname(next http.Handler) http.Handler {
 	})
 }
 
+func AdminOnlyMiddleware(next http.Handler) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if !IsAdmin(r) {
+			http.Error(w, "Unauthorized", http.StatusUnauthorized)
+			return
+		}
+
+		next.ServeHTTP(w, r)
+	})
+}
+
 func EnsureHostnameCosmosAPI(next http.Handler) http.Handler {
 	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 		og := GetMainConfig().HTTPConfig.Hostname